+++ frontend/src/api.js
@@ -428,12 +428,21 @@ class Api {
       subprojectId
     });
 
-  closeWorkflowItem = (projectId, subprojectId, workflowitemId) =>
-    instance.post(`/workflowitem.close`, {
+  closeWorkflowItem = (projectId, subprojectId, workflowitemId, rejectReason) => {
+    if(rejectReason === "")
+      return instance.post(`/workflowitem.close`, {
+        projectId,
+        subprojectId,
+        workflowitemId
+      });
+    else
+    return instance.post(`/workflowitem.close`, {
       projectId,
       subprojectId,
-      workflowitemId
+      workflowitemId,
+      rejectReason
     });
+  }
 
   fetchNotifications = (offset, limit) => {
     let url = removeEmptyQueryParams(`/notification.list?offset=${offset}&limit=${limit}`);