+++ frontend/src/languages/french.js
@@ -87,6 +87,7 @@ const fr = {
     not_assigned: "Non engagé",
     not_disbursed: "Non payé",
     not_ok: "Pas OK",
+    ok:"Ok",
     not_projected: "Non estimé",
     open: "Ouvert",
     organization: "Organisation",
@@ -98,6 +99,8 @@ const fr = {
     redacted: "Expurgé",
     reset: "Réinitialiser",
     revoke: "Révoquer",
+    reject:"Rejette",
+    rejected:" Rejeté",
     search: "Chercher",
     show_permissions: "Afficher les autorisations",
     status: "Statut",
@@ -243,8 +246,9 @@ const fr = {
     workflowitem_type: "Type de l’étape de workflow",
     workflowitem_type_general: "Créer une étape de workflow de type général.",
     workflowitem_type_restricted:
-      "Lors de l'attribution d'un workflow restreint, les autorisations sont automatiquement accordées et révoquées. Le cédant ne conservera que les autorisations de visualisation"
-  },
+      "Lors de l'attribution d'un workflow restreint, les autorisations sont automatiquement accordées et révoquées. Le cédant ne conservera que les autorisations de visualisation",
+    workflow_reject_reason:"Motif de rejet"
+    },
 
   snackbar: {
     update_succeed_message: "{0} modifiée",
@@ -353,7 +357,9 @@ const fr = {
     user_group: "Utilisateur/Groupe",
     workflowitem_close: "Fermer l’étape de workflow ",
     workflowitem_close_text: "Êtes-vous sûr de vouloir fermer cette étape de workflow?",
-    workflowitem_create: "créer des étapes de workflow"
+    workflowitem_create: "créer des étapes de workflow",
+    workflowitem_close_reject:"Rejet avec commentaire",
+    workflowitem_close_accept: "Acceptez"
   },
 
   intents: {