+++ frontend/src/pages/Confirmation/ConfirmationContainer.js
@@ -1,7 +1,6 @@
 import _isEmpty from "lodash/isEmpty";
 import React from "react";
 import { connect } from "react-redux";
-
 import { toJS } from "../../helper";
 import { fetchProjectPermissions, grantProjectPermission, revokeProjectPermission } from "../Overview/actions";
 import {
@@ -12,7 +11,7 @@ import {
   grantSubProjectPermission,
   revokeSubProjectPermission
 } from "../SubProjects/actions";
-import { fetchGroups, disableUser, enableUser, fetchUserAssignments, cleanUserAssignments } from "../Users/actions";
+import { cleanUserAssignments, disableUser, enableUser, fetchGroups, fetchUserAssignments } from "../Users/actions";
 import {
   assignSubproject,
   assignWorkflowItem,
@@ -21,18 +20,19 @@ import {
   createWorkflowItem,
   fetchWorkflowItemPermissions,
   grantWorkflowItemPermission,
-  revokeWorkflowItemPermission
+  revokeWorkflowItemPermission,
+  storeRejectReason
 } from "../Workflows/actions";
 import {
   additionalActionUpdateRequired,
   cancelConfirmation,
-  confirmConfirmation,
   closeConfirmation,
+  confirmConfirmation,
   executeConfirmedActions,
+  showValidationErrorMessage,
   storeAdditionalActions,
   storePostActions,
-  storeRequestedPermissions,
-  showValidationErrorMessage
+  storeRequestedPermissions
 } from "./actions";
 import ConfirmationDialog from "./ConfirmationDialog";
 import { applyOriginalActions, createAdditionalActions } from "./createAdditionalActions";
@@ -170,7 +170,9 @@ class ConfirmationContainer extends React.Component {
       executedOriginalActions,
       executingOriginalActions,
       originalActionsExecuted,
-      failedOriginalAction
+      failedOriginalAction,
+      storeRejectReason,
+      rejectReason
     } = this.props;
 
     if (confirmationDialogOpen) {
@@ -209,6 +211,8 @@ class ConfirmationContainer extends React.Component {
           executingOriginalActions={executingOriginalActions}
           originalActionsExecuted={originalActionsExecuted}
           failedOriginalAction={failedOriginalAction}
+          storeRejectReason={storeRejectReason}
+          rejectReason={rejectReason}
         />
       );
     } else {
@@ -296,11 +300,12 @@ const mapDispatchToProps = dispatch => {
     additionalActionUpdateRequired: required => dispatch(additionalActionUpdateRequired(required)),
     closeProject: pId => dispatch(closeProject(pId, true)),
     closeSubproject: (pId, sId) => dispatch(closeSubproject(pId, sId, true)),
-    closeWorkflowItem: (pId, sId, wId) => dispatch(closeWorkflowItem(pId, sId, wId, true)),
+    closeWorkflowItem: (pId, sId, wId, isRejected) => dispatch(closeWorkflowItem(pId, sId, wId, isRejected, true)),
     disableUser: userId => dispatch(disableUser(userId)),
     enableUser: userId => dispatch(enableUser(userId)),
     fetchUserAssignments: userId => dispatch(fetchUserAssignments(userId)),
-    cleanUserAssignments: () => dispatch(cleanUserAssignments())
+    cleanUserAssignments: () => dispatch(cleanUserAssignments()),
+    storeRejectReason: reason => dispatch(storeRejectReason(reason))
   };
 };
 
@@ -339,7 +344,8 @@ const mapStateToProps = state => {
     requestedPermissions: state.getIn(["confirmation", "requestedPermissions"]),
     isPayloadValidationFailed: state.getIn(["confirmation", "isPayloadValidationFailed"]),
     groups: state.getIn(["users", "groups"]),
-    userAssignments: state.getIn(["users", "userAssignments"])
+    userAssignments: state.getIn(["users", "userAssignments"]),
+    rejectReason: state.getIn(["workflow", "rejectReason"])
   };
 };